Repairs to Orders.  Use Projects instead of Orders to time progressed actions

+++ SpaceOpera/Core/Economics/Structure.cs
@@ -8,7 +8,7 @@ namespace SpaceOpera.Core.Economics
         public string Key { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
         public uint MaxWorkers { get; set; }
-        public uint BuildTime { get; set; }
+        public float BuildTime { get; set; }
         public MultiQuantity<IMaterial> Cost { get; set; } = new();
 
         public override string ToString()
